All done!
Finally, the game is complete. I have personally executed it and it was great. 
There was no error which is really a good thing. 
Next please check the movements (controls) otherwise it is ready to play.

And I think that  this project is now completed. 
It has the game, a readme and a contribution file too. So now we can start another project. 

Please reply soon so we can discuss about our next project. I will surely help you in whatever thing I can. Hope you will be happy by working with me.

Thanks!!